requirements.txt
python-telegram-bot==20.6
timezonefinder==6.5.9
numpy>=2.0
tzdata>=2024.1
psycopg2-binary